Revert to stock UI for now
